Changed matrix.Vector to be a type alias of blas64.Vector
This is a proof of concept and I would not dig deeply into the changes
immediately, because if you don’t like the premise then the rest is
irrelevant.

At first I was thinking about adding a BinaryMarshal and
BinaryUnmarshal to blas64.Vector to satisfy #337, but those methods are
supposed to be composable, so I thought maybe it would make sense to
add to blas64.Vector and then call from matrix.

But I’m really lazy and I’d rather not have to call it from matrix at
all.  By changing the type of matrix.Vector to blas64.Vector, it will
automatically pick up any methods from the original type.  As a bonus,
calling blas funcs on matrix vectors is a snap.  I imagine this
approach could be extended to other matrix types, and would allow
library callers to get down to a lower implementation level whenever
they want to take the gloves off.

(github autocorrect keeps trying to call this “composable" idea
“compostable”, hopefully you won’t feel the same…)